+++ projection/winrt-sdk/src/winrt_sdk/pywinrt/pybase.h
@@ -1812,6 +1812,261 @@ namespace py
         }
     };
 
+    template<typename K, typename V>
+    struct python_key_value_pair
+        : winrt::implements<
+              python_key_value_pair<K, V>,
+              winrt::Windows::Foundation::Collections::IKeyValuePair<K, V>>
+    {
+        K _key;
+        V _value;
+
+        python_key_value_pair(K key, V value) : _key(key), _value(value)
+        {
+        }
+
+        K Key() const
+        {
+            return _key;
+        }
+
+        V Value() const
+        {
+            return _value;
+        }
+    };
+
+    template<typename K, typename V>
+    struct python_mapping_iterator
+        : winrt::implements<
+              python_mapping_iterator<K, V>,
+              winrt::Windows::Foundation::Collections::IIterator<
+                  winrt::Windows::Foundation::Collections::IKeyValuePair<K, V>>>
+    {
+        using KVPair = winrt::Windows::Foundation::Collections::IKeyValuePair<K, V>;
+
+        pyobj_handle _mapping;
+        pyobj_handle _iterator;
+        std::optional<KVPair> _current_value;
+
+        static std::optional<KVPair> get_next(
+            pyobj_handle const& mapping, pyobj_handle const& iterator)
+        {
+            pyobj_handle next_key{PyIter_Next(iterator.get())};
+
+            if (!next_key)
+            {
+                if (PyErr_Occurred())
+                {
+                    throw python_exception();
+                }
+
+                return std::nullopt;
+            }
+
+            pyobj_handle next_value{PyObject_GetItem(mapping.get(), next_key.get())};
+
+            if (!next_value)
+            {
+                throw python_exception();
+            }
+
+            auto key = converter<K>::convert_to(next_key.get());
+            auto value = converter<V>::convert_to(next_value.get());
+
+            return winrt::make<python_key_value_pair<K, V>>(key, value);
+        }
+
+        python_mapping_iterator(PyObject* mapping) : _mapping(mapping)
+        {
+            Py_INCREF(_mapping.get());
+
+            _iterator = pyobj_handle{PyObject_GetIter(_mapping.get())};
+
+            if (!_iterator)
+            {
+                throw python_exception();
+            }
+
+            _current_value = get_next(_mapping, _iterator);
+        }
+
+        KVPair Current() const
+        {
+            return _current_value.value();
+        }
+
+        bool HasCurrent() const
+        {
+            return _current_value.has_value();
+        }
+
+        bool MoveNext()
+        {
+            _current_value = get_next(_mapping, _iterator);
+            return _current_value.has_value();
+        }
+
+        uint32_t GetMany(winrt::array_view<KVPair> /*unused*/)
+        {
+            // TODO: implement GetMany
+            PyErr_Format(
+                PyExc_NotImplementedError,
+                "py::python_mapping_iterator<%s, %s>::GetMany() is not implemented",
+                typeid(K).name(),
+                typeid(V).name());
+            throw python_exception();
+        }
+    };
+
+    template<typename K, typename V>
+    struct python_map
+        : winrt::implements<
+              python_map<K, V>,
+              winrt::Windows::Foundation::Collections::IMap<K, V>,
+              winrt::Windows::Foundation::Collections::IIterable<
+                  winrt::Windows::Foundation::Collections::IKeyValuePair<K, V>>>
+    {
+        pyobj_handle _mapping;
+
+        // It is up to the caller to ensure that this is a mapping object!
+        explicit python_map(PyObject* mapping) : _mapping(mapping)
+        {
+            Py_INCREF(_mapping.get());
+        }
+
+        void Clear()
+        {
+            pyobj_handle result{PyObject_CallMethod(_mapping.get(), "clear", nullptr)};
+
+            if (!result)
+            {
+                throw python_exception();
+            }
+        }
+
+        winrt::Windows::Foundation::Collections::IMapView<K, V> GetView() const
+        {
+            // TODO: implement GetView
+            PyErr_Format(
+                PyExc_NotImplementedError,
+                "py::python_map<%s, %s>::GetView() is not implemented",
+                typeid(K).name(),
+                typeid(V).name());
+            throw python_exception();
+        }
+
+        bool HasKey(K const& key) const
+        {
+            pyobj_handle py_key{converter<K>::convert(key)};
+
+            if (!py_key)
+            {
+                throw python_exception();
+            }
+
+            auto ret = PyMapping_HasKeyWithError(_mapping.get(), py_key.get());
+
+            if (ret == -1)
+            {
+                throw python_exception();
+            }
+
+            return static_cast<bool>(ret);
+        }
+
+        bool Insert(K const& key, V const& value)
+        {
+            pyobj_handle py_key{converter<K>::convert(key)};
+
+            if (!py_key)
+            {
+                throw python_exception();
+            }
+
+            pyobj_handle py_value{converter<V>::convert(value)};
+
+            if (!py_value)
+            {
+                throw python_exception();
+            }
+
+            auto result = HasKey(key);
+
+            if (PyObject_SetItem(_mapping.get(), py_key.get(), py_value.get()) == -1)
+            {
+                throw python_exception();
+            }
+
+            return result;
+        }
+
+        V Lookup(K const& key) const
+        {
+            pyobj_handle py_key{converter<K>::convert(key)};
+
+            if (!py_key)
+            {
+                throw python_exception();
+            }
+
+            pyobj_handle item{PyObject_GetItem(_mapping.get(), py_key.get())};
+
+            if (!item)
+            {
+                if (PyErr_ExceptionMatches(PyExc_KeyError))
+                {
+                    PyErr_Clear();
+                    throw winrt::hresult_out_of_bounds();
+                }
+
+                throw python_exception();
+            }
+
+            return converter<V>::convert_to(item.get());
+        }
+
+        void Remove(K const& key)
+        {
+            pyobj_handle py_key{converter<K>::convert(key)};
+
+            if (!py_key)
+            {
+                throw python_exception();
+            }
+
+            if (PyObject_DelItem(_mapping.get(), py_key.get()) == -1)
+            {
+                if (PyErr_ExceptionMatches(PyExc_KeyError))
+                {
+                    PyErr_Clear();
+                    throw winrt::hresult_out_of_bounds();
+                }
+
+                throw python_exception();
+            }
+        }
+
+        uint32_t Size() const
+        {
+            auto size = PyMapping_Size(_mapping.get());
+
+            if (size == -1 && PyErr_Occurred())
+            {
+                throw python_exception();
+            }
+
+            return static_cast<uint32_t>(size);
+        }
+
+        winrt::Windows::Foundation::Collections::IIterator<
+            winrt::Windows::Foundation::Collections::IKeyValuePair<K, V>>
+        First() const
+        {
+            return winrt::make<python_mapping_iterator<K, V>>(_mapping.get());
+        }
+    };
+
     template<typename T>
     std::optional<T> convert_interface_to(PyObject* obj)
     {
@@ -1852,8 +2107,6 @@ namespace py
                    obj)
             ->obj.as<T>();
     }
-
-    // TODO: specialization for Python Mapping Protocol -> IMap[View]
 
     template<typename TItem>
     struct converter<winrt::Windows::Foundation::Collections::IIterable<TItem>>
@@ -1944,6 +2197,33 @@ namespace py
         }
     };
 
+    template<typename TKey, typename TValue>
+    struct converter<winrt::Windows::Foundation::Collections::IMap<TKey, TValue>>
+    {
+        using TMap = winrt::Windows::Foundation::Collections::IMap<TKey, TValue>;
+
+        static PyObject* convert(TMap const& instance) noexcept
+        {
+            return wrap(instance);
+        }
+
+        static auto convert_to(PyObject* obj)
+        {
+            if (auto result = convert_interface_to<TMap>(obj))
+            {
+                return result.value();
+            }
+
+            if (!PyMapping_Check(obj))
+            {
+                PyErr_SetString(PyExc_TypeError, "expected a mapping");
+                throw python_exception();
+            }
+
+            return winrt::make<python_map<TKey, TValue>>(obj);
+        }
+    };
+
     template<typename T>
     struct is_specialized_interface : std::false_type
     {
@@ -1968,6 +2248,12 @@ namespace py
     template<typename TItem>
     struct is_specialized_interface<
         winrt::Windows::Foundation::Collections::IVector<TItem>> : std::true_type
+    {
+    };
+
+    template<typename TKey, typename TValue>
+    struct is_specialized_interface<
+        winrt::Windows::Foundation::Collections::IMap<TKey, TValue>> : std::true_type
     {
     };